+++ src/Illuminate/Foundation/Console/NotificationMakeCommand.php
@@ -7,6 +7,8 @@ use Symfony\Component\Console\Input\InputOption;
 
 class NotificationMakeCommand extends GeneratorCommand
 {
+    use MakeMailTrait;
+
     /**
      * The console command name.
      *
@@ -27,55 +29,6 @@ class NotificationMakeCommand extends GeneratorCommand
      * @var string
      */
     protected $type = 'Notification';
-
-    /**
-     * Execute the console command.
-     *
-     * @return void
-     */
-    public function handle()
-    {
-        if (parent::handle() === false && ! $this->option('force')) {
-            return;
-        }
-
-        if ($this->option('markdown')) {
-            $this->writeMarkdownTemplate();
-        }
-    }
-
-    /**
-     * Write the Markdown template for the mailable.
-     *
-     * @return void
-     */
-    protected function writeMarkdownTemplate()
-    {
-        $path = resource_path('views/'.str_replace('.', '/', $this->option('markdown'))).'.blade.php';
-
-        if (! $this->files->isDirectory(dirname($path))) {
-            $this->files->makeDirectory(dirname($path), 0755, true);
-        }
-
-        $this->files->put($path, file_get_contents(__DIR__.'/stubs/markdown.stub'));
-    }
-
-    /**
-     * Build the class with the given name.
-     *
-     * @param  string  $name
-     * @return string
-     */
-    protected function buildClass($name)
-    {
-        $class = parent::buildClass($name);
-
-        if ($this->option('markdown')) {
-            $class = str_replace('DummyView', $this->option('markdown'), $class);
-        }
-
-        return $class;
-    }
 
     /**
      * Get the stub file for the generator.
